Minimize listen time to reduce power usage
This involved using a multimeter (surprisingly my fluke didn't have \muamp on it) to measure the current going to the nRf24L01 module (through the adapter with 5v -> 3.3v).

Nominal results went from about 17mA on full nrf.listen to about 3mA with the new timing within receivePayload (idle is about 2mA with the LED unsoldered for reference). I think this makes it less likely to 'catch' the transmit window but I wasn't have any problems so far. In the future I expect that with additional non-nRf24 processing time I may have to increase the window or figure something else out (to be honest the receive side is planned to be plugged in, so leaving nrf.listen on all the time may be fine).